+++ src/mlpack/core/tree/binary_space_tree/traits.hpp
@@ -21,7 +21,7 @@ namespace tree {
 template<typename MetricType,
          typename StatisticType,
          typename MatType,
-         template<typename BoundMetricType> class BoundType,
+         template<typename BoundMetricType, typename...> class BoundType,
          template<typename SplitBoundType, typename SplitMatType>
              class SplitType>
 class TreeTraits<BinarySpaceTree<MetricType, StatisticType, MatType, BoundType,