+++ src/mlpack/tests/activation_functions_test.cpp
@@ -1,6 +1,7 @@
 /**
  * @file activation_functions_test.cpp
  * @author Marcus Edel
+ * @author Dhawal Arora
  *
  * Tests for the various activation functions.
  */
@@ -21,6 +22,8 @@
 #include <mlpack/methods/ann/layer/linear_layer.hpp>
 #include <mlpack/methods/ann/layer/base_layer.hpp>
 #include <mlpack/methods/ann/layer/binary_classification_layer.hpp>
+#include <mlpack/methods/ann/layer/leaky_relu_layer.hpp>
+#include <mlpack/methods/ann/layer/hard_tanh_layer.hpp>
 
 #include <boost/test/unit_test.hpp>
 #include "old_boost_test_definitions.hpp"
@@ -120,6 +123,98 @@ void CheckInverseCorrect(const arma::colvec input)
   }
 }
 
+/*
+ * Implementation of the HardTanH activation function test. The function is
+ * implemented as a HardTanH Layer in hard_tanh_layer.hpp
+ *
+ * @param input Input data used for evaluating the HardTanH activation function.
+ * @param target Target data used to evaluate the HardTanH activation.
+ */
+void CheckHardTanHActivationCorrect(const arma::colvec input,
+                                    const arma::colvec target)
+{
+  HardTanHLayer<> htf;
+
+  // Test the activation function using the entire vector as input.
+  arma::colvec activations;
+  htf.Forward(input, activations);
+  for (size_t i = 0; i < activations.n_elem; i++)
+  {
+    BOOST_REQUIRE_CLOSE(activations.at(i), target.at(i), 1e-3);
+  }
+}
+
+/*
+ * Implementation of the HardTanH activation function derivative test. The
+ * derivative is implemented as HardTanH Layer in hard_tanh_layer.hpp
+ *
+ * @param input Input data used for evaluating the HardTanH activation function.
+ * @param target Target data used to evaluate the HardTanH activation.
+ */
+void CheckHardTanHDerivativeCorrect(const arma::colvec input,
+                                    const arma::colvec target)
+{
+  HardTanHLayer<> htf;
+
+  // Test the calculation of the derivatives using the entire vector as input.
+  arma::colvec derivatives;
+
+  // This error vector will be set to 1 to get the derivatives.
+  arma::colvec error(input.n_elem);
+  htf.Backward(input, (arma::colvec)error.ones(), derivatives);
+  for (size_t i = 0; i < derivatives.n_elem; i++)
+  {
+    BOOST_REQUIRE_CLOSE(derivatives.at(i), target.at(i), 1e-3);
+  }
+}
+
+/*
+ * Implementation of the LeakyReLU activation function test. The function is  
+ * implemented as LeakyReLU layer in the file leaky_relu_layer.hpp
+ * 
+ * @param input Input data used for evaluating the LeakyReLU activation function.
+ * @param target Target data used to evaluate the LeakyReLU activation.
+ */
+void CheckLeakyReLUActivationCorrect(const arma::colvec input,
+                                     const arma::colvec target)
+{
+  LeakyReLULayer<> lrf;
+
+  // Test the activation function using the entire vector as input.
+  arma::colvec activations;
+  lrf.Forward(input, activations);
+  for (size_t i = 0; i < activations.n_elem; i++)
+  {
+    BOOST_REQUIRE_CLOSE(activations.at(i), target.at(i), 1e-3);
+  }
+}
+
+/*
+ * Implementation of the LeakyReLU activation function derivative test. 
+ * The derivative function is implemented as LeakyReLU layer in the file 
+ * leaky_relu_layer.hpp
+ *
+ * @param input Input data used for evaluating the LeakyReLU activation function.
+ * @param target Target data used to evaluate the LeakyReLU activation.
+ */
+
+void CheckLeakyReLUDerivativeCorrect(const arma::colvec input, 
+                                     const arma::colvec target)
+{
+  LeakyReLULayer<> lrf;
+
+  // Test the calculation of the derivatives using the entire vector as input.
+  arma::colvec derivatives;
+
+  // This error vector will be set to 1 to get the derivatives.
+  arma::colvec error(input.n_elem);
+  lrf.Backward(input, (arma::colvec)error.ones(), derivatives);
+  for (size_t i = 0; i < derivatives.n_elem; i++)
+  {
+    BOOST_REQUIRE_CLOSE(derivatives.at(i), target.at(i), 1e-3);
+  }
+}
+
 /**
  * Basic test of the tanh function.
  */
@@ -199,4 +294,35 @@ BOOST_AUTO_TEST_CASE(RectifierFunctionTest)
       desiredDerivatives);
 }
 
+/**
+ * Basic test of the LeakyReLU function.
+ */
+BOOST_AUTO_TEST_CASE(LeakyReLUFunctionTest)
+{
+  const arma::colvec desiredActivations("-0.06 3.2 4.5 -3.006 \
+                                         1 -0.03 2 0");
+
+  const arma::colvec desiredDerivatives("0.03 1 1 0.03 \
+                                         1 0.03 1 1");
+
+  CheckLeakyReLUActivationCorrect(activationData, desiredActivations);
+  CheckLeakyReLUDerivativeCorrect(desiredActivations, desiredDerivatives);
+}
+
+/**
+ * Basic test of the HardTanH function.
+ */
+BOOST_AUTO_TEST_CASE(HardTanHFunctionTest)
+{
+  const arma::colvec desiredActivations("-1 1 1 -1 \
+                                         1 -1 1 0");
+
+  const arma::colvec desiredDerivatives("0 0 0 0 \
+                                         1 1 0 1");
+
+  CheckHardTanHActivationCorrect(activationData, desiredActivations);
+  CheckHardTanHDerivativeCorrect(activationData, desiredDerivatives);
+}
+
 BOOST_AUTO_TEST_SUITE_END();
+