+++ src/mlpack/tests/hoeffding_tree_test.cpp
@@ -902,9 +902,10 @@ BOOST_AUTO_TEST_CASE(BatchTrainingTest)
   arma::mat spiralDataset(2, 10000);
   for (size_t i = 0; i < 10000; ++i)
   {
-    // One circle every 2000 samples.
-    const double magnitude = 2.0 + (double(i) / 20000.0);
-    const double angle = (i % 20000) * (2 * M_PI);
+    // One circle every 20000 samples.  Plus some noise.
+    const double magnitude = 2.0 + (double(i) / 20000.0) +
+        0.5 * mlpack::math::Random();
+    const double angle = (i % 20000) * (2 * M_PI) + mlpack::math::Random();
 
     const double x = magnitude * cos(angle);
     const double y = magnitude * sin(angle);
@@ -936,52 +937,39 @@ BOOST_AUTO_TEST_CASE(BatchTrainingTest)
     l[i] = labels[indices[i]];
   }
 
+  // Split into a training set and a test set.
+  arma::mat trainingData = d.cols(0, 4999);
+  arma::mat testData = d.cols(5000, 9999);
+  arma::Row<size_t> trainingLabels = l.subvec(0, 4999);
+  arma::Row<size_t> testLabels = l.subvec(5000, 9999);
+
   data::DatasetInfo info(2);
 
   // Now build two decision trees; one in batch mode, and one in streaming mode.
   // We need to set the confidence pretty high so that the streaming tree isn't
   // able to have enough samples to build to the same leaves.
-  HoeffdingTree<> batchTree(d, info, l, 5, true, 0.999);
-  HoeffdingTree<> streamTree(d, info, l, 5, false, 0.999);
-
-  size_t batchNodes = 0, streamNodes = 0;
-  std::stack<HoeffdingTree<>*> queue;
-  queue.push(&batchTree);
-  while (!queue.empty())
-  {
-    ++batchNodes;
-    HoeffdingTree<>* node = queue.top();
-    queue.pop();
-    for (size_t i = 0; i < node->NumChildren(); ++i)
-      queue.push(&node->Child(i));
-  }
-  queue.push(&streamTree);
-  while (!queue.empty())
-  {
-    ++streamNodes;
-    HoeffdingTree<>* node = queue.top();
-    queue.pop();
-    for (size_t i = 0; i < node->NumChildren(); ++i)
-      queue.push(&node->Child(i));
-  }
+  HoeffdingTree<> batchTree(trainingData, info, trainingLabels, 5, true,
+      0.99999999);
+  HoeffdingTree<> streamTree(trainingLabels, info, trainingLabels, 5, false,
+      0.99999999);
 
   // Ensure that the performance of the batch tree is better.
   size_t batchCorrect = 0;
   size_t streamCorrect = 0;
-  for (size_t i = 0; i < 10000; ++i)
+  for (size_t i = 0; i < 5000; ++i)
   {
-    size_t streamLabel = streamTree.Classify(spiralDataset.col(i));
-    size_t batchLabel = batchTree.Classify(spiralDataset.col(i));
+    size_t streamLabel = streamTree.Classify(testData.col(i));
+    size_t batchLabel = batchTree.Classify(testData.col(i));
 
-    if (streamLabel == labels[i])
+    if (streamLabel == testLabels[i])
       ++streamCorrect;
-    if (batchLabel == labels[i])
+    if (batchLabel == testLabels[i])
       ++batchCorrect;
   }
 
   // The batch tree must be a bit better than the stream tree.  But not too
   // much, since the accuracy is already going to be very high.
-  BOOST_REQUIRE_GT(batchCorrect, streamCorrect);
+  BOOST_REQUIRE_GE(batchCorrect, streamCorrect);
 }
 
 // Make sure that changing the confidence properly propagates to all leaves.